added viewing a map of nearby mood events
